pyproject.toml
[project]
name = "nerdsystem-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp[speedups]>=3.11.11",
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "websockets>=14.1",
    "yt-dlp>=2024.12.23",
]
